# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Secure Patrol API
  version: 1.0.0
  description: |
    Dokumentasi API Secure Patrol. **Hanya tersedia di environment development** (`APP_ENV=development`).

    ## Cara mencoba API di halaman ini
    1. Klik **Authorize**, isi **AppId** (`X-App-Id`) dan **AppKey**. App Key hanya dipakai browser untuk
       menandatangani request dan **tidak pernah dikirim** ke server.
    2. Panggil `POST /auth/login`. Access token otomatis dipasang ke **BearerAuth**.
    3. Endpoint lain siap dicoba.

    Belum punya App ID/Key? Jalankan `go run . create-app-client -name "Dev Docs" -platform web`.

    ## Keamanan request
    Setiap request ke `/api/v1` wajib ditandatangani dengan header `X-App-Id`, `X-Timestamp`, `X-Nonce`, dan
    `X-Signature`. Halaman ini membuat ketiga header terakhir secara otomatis.

    ```
    payload   = METHOD \n REQUEST_URI \n TIMESTAMP \n NONCE \n SHA256_HEX(BODY)
    signature = HEX(HMAC_SHA256(APP_KEY, payload))
    ```

    Detail lengkap dan contoh kode mobile ada di `docs/AUTH.md`.

    ## Pembatasan platform
    **Login per platform:** App Client `android` dan `ios` bisa dipakai semua role. App Client `web` (dan `server`)
    hanya untuk Super-Admin, Manager Keamanan, Kepala Keamanan, dan Admin Keamanan; role lain mendapat `403`
    `your role is not allowed to sign in on this platform` saat login, refresh, maupun request berikutnya.

...
    membuat, mengubah, dan menghapus patrol point dan patrol shift.

...

...

...
servers:
...
security:
  - AppId: []
    AppKey: []
    BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Tukar refresh token dengan pasangan token baru
      description: |
        Refresh token hanya bisa dipakai sekali. Simpan refresh token baru dari response.
        Memakai refresh token lama akan mematikan seluruh session.
      security:
        - AppId: []
          AppKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token baru
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /auth/logout-all:
    post:
      operationId: logoutAll
      tags: [Auth]
      summary: Logout dari semua perangkat
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/change-password:
    put:
      operationId: changePassword
      tags: [Auth]
      summary: Ganti password sendiri
      description: Semua perangkat lain otomatis logout; perangkat ini tetap login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /app-clients:
    get:
      operationId: listAppClients
      tags: [App Clients]
      summary: Daftar app client
      description: '**Hanya platform web.**'
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Daftar app client
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/AppClient' }
                          pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createAppClient
      tags: [App Clients]
      summary: Buat app client (App ID + App Key)
      description: '**Hanya platform web.** App Key hanya ditampilkan sekali di response ini.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppClientRequest'
      responses:
        '201':
          description: App client dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppClientCredentialResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /app-clients/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getAppClient
      tags: [App Clients]
      summary: Detail app client
      description: '**Hanya platform web.**'
      responses:
        '200':
          description: Detail app client (tanpa App Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppClientResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateAppClient
      tags: [App Clients]
      summary: Ubah app client
      description: '**Hanya platform web.** App client yang sedang dipakai untuk request ini tidak bisa dinonaktifkan.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppClientRequest'
      responses:
        '200':
          description: App client diubah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppClientResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      operationId: deleteAppClient
      tags: [App Clients]
      summary: Hapus app client (soft delete)
      description: '**Hanya platform web.** App client yang sedang dipakai untuk request ini tidak bisa dihapus.'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    AppId:
      type: apiKey
      in: header
      name: X-App-Id
      description: App ID aplikasi, contoh `sp_web_k3m9qx2ht5vbn4rw_w8xq2a`.
    AppKey:
      type: apiKey
      in: header
      name: X-App-Key
      description: |
        Khusus halaman dokumentasi. App Key dipakai browser untuk membuat `X-Signature`
        lalu dihapus dari request, sehingga tidak pernah terkirim ke server.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token dari `/auth/login`. Terisi otomatis setelah login di halaman ini.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    Search:
      name: search
      in: query
      schema: { type: string }
...
  responses:
...
    Success:
      description: Berhasil
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
...
    Unauthorized:
      description: App signature atau token tidak valid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Role tidak memiliki akses
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Data tidak ditemukan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ValidationError:
      description: Input tidak valid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    Meta:
      type: object
      properties:
        message: { type: string, example: Login success }
        code: { type: integer, example: 200 }
        status: { type: string, example: success }
    Envelope:
      type: object
      properties:
        meta: { $ref: '#/components/schemas/Meta' }
        data: {}
    ErrorResponse:
      type: object
      properties:
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
          example: { message: Validation error, code: 422, status: Error }
        data:
          description: Detail error (string, daftar pesan validasi, atau null)
          oneOf:
            - type: string
            - type: array
              items: { type: string }
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 10 }
        total: { type: integer, example: 42 }
        total_pages: { type: integer, example: 5 }
...
    User:
      type: object
      properties:
        id: { type: integer, format: int64, example: 6 }
        name: { type: string, example: Budi Santoso }
        email: { type: string, format: email, example: budi@securepatrol.local }
        role:
          type: object
          properties:
            id: { type: integer, format: int64, example: 5 }
            code: { type: string, example: security_team }
            name: { type: string, example: Tim Keamanan }
        is_active: { type: boolean, example: true }
        is_locked: { type: boolean, example: false, description: Terkunci karena terlalu banyak salah password }
        face_photo_url: { type: string, nullable: true, example: /api/v1/users/6/face-photo }
        face_photo_updated_at:
          type: string
          format: date-time
          nullable: true
          description: Berubah setiap foto wajah diganti. Aplikasi mobile mengunduh ulang foto jika nilainya berbeda dari cache.
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    ChangePasswordRequest:
      type: object
      required: [old_password, password, password_confirmation]
      properties:
        old_password: { type: string, format: password }
        password: { type: string, format: password }
        password_confirmation: { type: string, format: password }
    Token:
      type: object
      properties:
        token_type: { type: string, example: Bearer }
        access_token: { type: string }
        expires_in: { type: integer, description: Detik sampai access token kedaluwarsa, example: 3600 }
        expires_at: { type: string, format: date-time }
        refresh_token: { type: string, description: Hanya bisa dipakai sekali }
        refresh_expires_at: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/LoginUser' }
        config: { $ref: '#/components/schemas/AppConfig' }
        settings: { $ref: '#/components/schemas/AppSettings' }
    AppConfig:
      type: object
      description: Pengaturan untuk aplikasi mobile (validasi sebelum scan dan mode offline).
      properties:
        server_time:
          type: string
          format: date-time
          description: 'Waktu server. Pakai untuk mengoreksi jam HP: request ditolak jika jam HP selisih lebih dari request_timestamp_tolerance_seconds.'
        timezone: { type: string, example: Asia/Jakarta }
        request_timestamp_tolerance_seconds: { type: integer, example: 300 }
        location_radius_meters: { type: number, example: 100, description: Radius maksimal untuk titik yang mewajibkan validasi lokasi }
        face_match_min_score: { type: number, nullable: true, example: 0.8, description: 'null = server hanya mengecek face_verified' }
        max_offline_hours: { type: integer, example: 24 }
        max_scan_photos: { type: integer, example: 3 }
        max_photo_size_bytes: { type: integer, example: 5242880 }
    LoginUser:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            face_photo_base64:
              type: string
              format: byte
              nullable: true
              description: Foto wajah (base64 standar). Hanya diisi saat login; null jika user belum punya foto.
            face_photo_mime_type:
              type: string
              nullable: true
              enum: [image/jpeg, image/png]
              example: image/jpeg
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Token' }
    AppClient:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: Secure Patrol Android }
        platform: { type: string, enum: [android, ios, web, server] }
        app_id: { type: string, example: sp_and_k3m9qx2ht5vbn4rw_w8xq2a }
        key_hint: { type: string, example: '****s0yM' }
        description: { type: string }
        is_active: { type: boolean }
        expires_at: { type: string, format: date-time, nullable: true }
        last_used_at: { type: string, format: date-time, nullable: true }
        key_rotated_at: { type: string, format: date-time, nullable: true }
        previous_key_expires_at:
          type: string
          format: date-time
          nullable: true
          description: Batas waktu key lama masih diterima setelah rotasi
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AppClientResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/AppClient' }
    AppClientCredentialResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/AppClient'
                - type: object
                  properties:
                    app_key: { type: string, example: spk_-e2NwI_EJysxpO6hWXyKRxB4CWHpSDxm18PngULs0yM }
                    warning: { type: string }
    CreateAppClientRequest:
      type: object
      required: [name, platform]
      properties:
        name: { type: string, maxLength: 100 }
        platform: { type: string, enum: [android, ios, web, server] }
        description: { type: string, maxLength: 255 }
        expires_at: { type: string, format: date-time, nullable: true }
    UpdateAppClientRequest:
      type: object
      required: [name, is_active]
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 255 }
        is_active: { type: boolean }
        expires_at: { type: string, format: date-time, nullable: true }
...
    AppSettings:
      type: object
      description: Setting yang dikirim ke aplikasi pada login, refresh, dan GET /auth/me. Diatur di /settings.
      properties:
        patrol_location_radius_meters: { type: number, example: 100, description: Jarak maksimal HP ke titik patroli saat scan (meter) }
        patrol_max_offline_hours: { type: integer, example: 24, description: Batas umur scan offline yang masih diterima (jam) }
        face_mobile_accuracy: { type: number, example: 0.75, description: Skor kecocokan wajah (0-1) yang dibutuhkan HP untuk menyatakan wajah cocok }
        login_max_failed_attempts: { type: integer, example: 3, description: Jumlah salah password sebelum akun dikunci }
        login_lock_minutes: { type: integer, example: 5, description: Lama akun dikunci (menit) }
        access_token_ttl_minutes: { type: integer, example: 60, description: Masa berlaku access token (menit) }
        refresh_token_ttl_days: { type: integer, example: 30, description: Berapa lama tetap login tanpa memasukkan password (hari) }
...